requirements.txt
Django==4.2.8
djangorestframework==3.14.0
PyMySQL==1.1.0
mysqlclient==2.2.0
Pillow==10.1.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
jieba==0.42.1
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
python-dotenv==1.0.0
channels==4.0.0
channels-redis==4.2.0
websockets==12.0
daphne==4.0.0
